intermission: set black background behind all loading images
We add a black background behind the loading images displayed on the
SceneLoader and FejdStartup screens, in order to hide whatever is behind
them when the loading images do not scale up to the whole screen (e.g.
displaying a 16:9 loading image on a 21:9 screen).

In particular, on the SceneLoader screen the default Unity scene was
displayed, and on the FejdStartup screen the main menu was displayed
(leading to a visual discrepancy when the Hud screen was taking over).